+++ source/js/img-filter-handlers.js
@@ -1,6 +1,5 @@
-import { getThumbnailsMarkup, removePictureNodes } from './img-thumbnails.js';
+import { getThumbnailsMarkup } from './img-thumbnails.js';
 import {shufflePhotos} from './utils.js';
-import { openFullImageModal } from './full-img-modal.js';
 /* global _:readonly */
 const RERENDER_DELAY = 2000;
 const PHOTOS_COUNT = 25;
@@ -22,32 +21,17 @@ const filters = {
   }),
 }
 
-const onThumbnailsContainerMouseover = (evt) => {
-  if (evt.target && evt.target.closest('.picture__info')) {
-    evt.target.style.pointerEvents = 'none';
-  }
-}
-
-const onThumbnailsContainerClick = (evt) => {
-  const targetPicture = evt.target.closest('.picture');
-  if (evt.target && targetPicture) {
-    evt.preventDefault();
-    const currentImage = parseInt(evt.target.dataset.number);
-    const photo = originalPhotos[currentImage];
-    openFullImageModal(photo);
-    thumbnailsContainer.removeEventListener('click', onThumbnailsContainerClick);
-  }
+const renderPhotosMarkup = (photos) => {
+  const thumbnailsMarkup = getThumbnailsMarkup(photos);
+  thumbnailsContainer.appendChild(thumbnailsMarkup);
 }
 
 const applyPhotosFilter = _.debounce((evt) => {
   const button = evt.target.closest('.img-filters__button');
   const filterType = button.id;
   const filter = filters[filterType];
-  originalPhotos = filter();
-  const thumbnailsMarkup = getThumbnailsMarkup(originalPhotos);
-  thumbnailsContainer.appendChild(thumbnailsMarkup);
-  thumbnailsContainer.addEventListener('click', onThumbnailsContainerClick);
-  thumbnailsContainer.addEventListener('mouseover', onThumbnailsContainerMouseover);
+  const filteredPhotos = filter();
+  renderPhotosMarkup(filteredPhotos);
 },
 RERENDER_DELAY);
 
@@ -57,13 +41,11 @@ const onImgFiltersButtonsContainerClick = (evt) => {
   activeButton.classList.remove('img-filters__button--active');
   button.classList.add('img-filters__button--active');
   applyPhotosFilter(evt);
-  removePictureNodes();
 }
 
 export {
   applyPhotosFilter,
   setOriginalPhotos,
   onImgFiltersButtonsContainerClick,
-  onThumbnailsContainerClick,
-  onThumbnailsContainerMouseover
+  renderPhotosMarkup
 }
